curl_upload {curl}	R Documentation

Upload a File

Description

Upload a file to an http://, ftp://, or sftp:// (ssh) server. Uploading to HTTP means performing an HTTP PUT on that URL. Be aware that sftp is only available for libcurl clients built with libssh2.

Usage

curl_upload(file, url, verbose = TRUE, reuse = TRUE, ...)

Arguments

file
connection object or path to an existing file on disk

url
where to upload, should start with e.g. ftp://

verbose
emit some progress output

reuse
try to keep alive and recycle connections when possible

...
other arguments passed to handle_setopt, for example a username and password.

Examples

## Not run: # Upload package to winbuilder:
curl_upload('mypkg_1.3.tar.gz', 'ftp://win-builder.r-project.org/R-devel/')

## End(Not run)

[Package curl version 5.1.0 Index]